reoranize tbd files into module
