+++ experiments/opencog/pattern_matcher_vqa/pattern_matcher_vqa.py
@@ -12,7 +12,6 @@ import network_runner
 from opencog.atomspace import TruthValue
 from opencog.type_constructors import *
 from opencog.scheme_wrapper import *
-from opencog import bindlink
 
 from util import *
 from interface import FeatureExtractor, AnswerHandler, NoModelException
@@ -472,71 +471,6 @@ class PatternMatcherVqaPipeline:
             except BaseException as e:
                 logger.exception('Unexpected exception %s', e)
                 continue
-
-
-class TBD(PatternMatcherVqaPipeline):
-
-
-    def answerByPrograms(self, tdb_net, features, programs):
-        batch_size = features.size(0)
-        feat_input_volume = tdb_net.stem(features)
-        results = []
-        for n in range(batch_size):
-            feat_input = feat_input_volume[n:n + 1]
-            output = feat_input
-            program = []
-            for i in reversed(programs.data[n].cpu().numpy()):
-                module_type = tdb_net.vocab['program_idx_to_token'][i]
-                if module_type == '<NULL>':
-                    continue
-                program.append(module_type)
-            result = self.run_program(output, program)
-            results.append(result)
-        return results
-
-    def argmax(self, answer_set):
-        import tbd_helpers
-        items = []
-        key_attention, key_scene, key_shape_attention, key_shape_scene = tbd_helpers.generate_keys(self.atomspace)
-        for list_link in answer_set.get_out():
-            atoms = list_link.get_out()
-            value = -1
-            concept = None
-            for atom in atoms:
-                if atom.name.startswith("Data-"):
-                    value = tbd_helpers.extract_tensor(atom, key_attention, key_shape_attention).numpy().sum()
-                elif atom.name.startswith("BoundingBox"):
-                    continue
-                else:
-                    concept = atom.name
-            assert concept
-            assert value != -1
-            items.append((value, concept))
-        if not items:
-            return None
-        items.sort(reverse=True)
-        return items[0][1]
-
-    def run_program(self, features, program):
-        self.atomspace = pushAtomspace(self.atomspace)
-        self._add_scene_atom(features)
-        import tbd_helpers
-        eval_link, left, inheritance_set = tbd_helpers.return_prog(commands=tuple(reversed(program)), atomspace=self.atomspace)
-        bind_link = tbd_helpers.build_bind_link(self.atomspace, eval_link, inheritance_set)
-        result = bindlink.bindlink(self.atomspace, bind_link)
-        answer = self.argmax(result)
-        self.atomspace = popAtomspace(self.atomspace)
-        return answer
-
-    def _add_scene_atom(self, features):
-        import tbd_helpers
-        _, key_scene, _, key_shape_scene = tbd_helpers.generate_keys(self.atomspace)
-        data = FloatValue(list(features.numpy().flatten()))
-        boundingBoxInstance = self.atomspace.add_node(types.ConceptNode, 'BoundingBox1')
-        boundingBoxInstance.set_value(key_scene, data)
-        boundingBoxInstance.set_value(key_shape_scene, FloatValue(list(features.numpy().shape)))
-        box_concept = self.atomspace.add_node(types.ConceptNode, 'BoundingBox')
-        self.atomspace.add_link(types.InheritanceLink, [boundingBoxInstance, box_concept])
 
 
 ### MAIN
