+++ experiments/opencog/pattern_matcher_vqa/tbd_cog/tbd_helpers.py
