+++ experiments/opencog/pattern_matcher_vqa/tbd_cog/tbdas.scm
